requirement.txt
bctpy==0.6.1
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scipy==1.13.0
six==1.16.0
tzdata==2024.1
